refactor: Move component schemas into component file
