--- common/components/Card/CardContent/CardContent.schema.ts
@@ -1,3 +1,0 @@
-export interface CardContentPropSchema {
-  children?: React.ReactNode;
-}
