--- common/components/Card/CardFooter/CardFooter.schema.ts
@@ -1,12 +1,0 @@
-import { Color } from '../../../constants/Color';
-import { ButtonPropSchema } from '../../Button/Button.schema';
-
-export interface CardFooterActionsSchema
-  extends Omit<ButtonPropSchema, 'children'> {
-  text: string;
-}
-
-export interface CardFooterPropSchema {
-  buttonColor?: Color;
-  actions: CardFooterActionsSchema[];
-}
