--- common/components/Card/CardHeader/CardHeader.schema.ts
@@ -1,9 +1,0 @@
-import { Color } from '../../../constants/Color';
-
-export interface CardHeaderPropSchema {
-  title: string;
-  image?: string;
-  subTitle?: string;
-  actions?: any[];
-  actionsButtonColor?: Color;
-}
