--- common/components/Card/CardMedia/CardMedia.schema.ts
@@ -1,4 +1,0 @@
-export interface CardMediaPropSchema {
-  alt: string;
-  image: string;
-}
