--- common/components/Link/Link.schema.ts
@@ -1,10 +1,0 @@
-import React from 'react';
-import { UrlObject } from 'url';
-import { Color } from '../../constants/Color';
-
-export interface LinkPropSchema {
-  children: React.ReactNode;
-  href: string | UrlObject;
-  disabled?: boolean;
-  color?: Color;
-}
